+++ openmls/src/group/public_group/tests.rs
@@ -98,6 +98,10 @@ fn public_group() {
                 .merge_commit(public_provider.storage(), *staged_commit)
                 .unwrap()
         }
+        #[cfg(feature = "extensions-draft")]
+        ProcessedMessageContent::UnresolvedAppDataCommit(_) => {
+            panic!("Unexpected message type.")
+        }
     };
 
     let welcome: MlsMessageIn = welcome.into();
@@ -206,6 +210,10 @@ fn public_group() {
         | ProcessedMessageContent::ExternalJoinProposalMessage(_)
         | ProcessedMessageContent::StagedCommitMessage(_)
         | ProcessedMessageContent::OwnPendingCommit => panic!("Unexpected message type."),
+        #[cfg(feature = "extensions-draft")]
+        ProcessedMessageContent::UnresolvedAppDataCommit(_) => {
+            panic!("Unexpected message type.")
+        }
         ProcessedMessageContent::ProposalMessage(p) => {
             match p.proposal() {
                 Proposal::Remove(r) => assert_eq!(r.removed(), LeafNodeIndex::new(1)),
@@ -330,6 +338,10 @@ fn extract_staged_commit(ppm: ProcessedMessage) -> StagedCommit {
             panic!("Unexpected message type.")
         }
         ProcessedMessageContent::StagedCommitMessage(staged_content) => *staged_content,
+        #[cfg(feature = "extensions-draft")]
+        ProcessedMessageContent::UnresolvedAppDataCommit(_) => {
+            panic!("Unexpected message type.")
+        }
     }
 }